+++ src/ru/yandex/practicum/test/EpicEndpointsTest.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.tests;
+package ru.yandex.practicum.test;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ru.yandex.practicum.http.HttpTaskServer;
 import ru.yandex.practicum.tasks.Epic;
-import ru.yandex.practicum.tasks.Subtask;
 import ru.yandex.practicum.utils.Utils;
 
 import java.io.IOException;
@@ -18,12 +17,10 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.time.LocalDateTime;
 
-import static java.time.Month.AUGUST;
-
-public class PrioritizedEndpointsTest {
+public class EpicEndpointsTest {
     private HttpTaskServer httpTaskServer;
     private HttpClient client;
-    private static final String DEFAULT_PRIORITIZED_URI = "http://localhost:8080/prioritized";
+    private static final String DEFAULT_EPICS_URI = "http://localhost:8080/epics";
     private static final Gson gson = new GsonBuilder()
             .registerTypeAdapter(LocalDateTime.class, new Utils.LocalDateTimeAdapter())
             .create();
@@ -41,8 +38,8 @@ public class PrioritizedEndpointsTest {
     }
 
     @Test
-    public void getEmptyPrioritizedListTest() throws IOException, InterruptedException {
-        URI uri = URI.create(DEFAULT_PRIORITIZED_URI);
+    public void getEmptyListEpicsTest() throws IOException, InterruptedException {
+        URI uri = URI.create(DEFAULT_EPICS_URI);
         HttpRequest request = HttpRequest.newBuilder()
                 .GET()
                 .uri(uri)
@@ -52,8 +49,8 @@ public class PrioritizedEndpointsTest {
     }
 
     @Test
-    public void getPrioritizedListTest() throws IOException, InterruptedException {
-        URI uri1 = URI.create("http://localhost:8080/epics");
+    public void addEpicAndGetItTest() throws IOException, InterruptedException {
+        URI uri1 = URI.create(DEFAULT_EPICS_URI);
         Epic epic1 = new Epic("Съездить в Москву", "обязательно до лета");
         HttpRequest request1 = HttpRequest.newBuilder()
                 .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(epic1)))
@@ -63,32 +60,34 @@ public class PrioritizedEndpointsTest {
         HttpResponse<String> response1 = client.send(request1, HttpResponse.BodyHandlers.ofString());
         Assertions.assertEquals(200, response1.statusCode(), "Not right response");
 
-        URI uri2 = URI.create("http://localhost:8080/subtasks");
-        Subtask subtask11 = new Subtask("купить билеты", "дешёвые билеты", 1,
-                LocalDateTime.of(2023, AUGUST, 28, 13, 0), 60);
+        URI uri2 = URI.create(DEFAULT_EPICS_URI);
         HttpRequest request2 = HttpRequest.newBuilder()
-                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subtask11)))
-                .header("Content-Type", "application/json")
+                .GET()
                 .uri(uri2)
                 .build();
         HttpResponse<String> response2 = client.send(request2, HttpResponse.BodyHandlers.ofString());
         Assertions.assertEquals(200, response2.statusCode(), "Not right response");
+    }
 
-
-        URI uri3 = URI.create("http://localhost:8080/subtasks");
-        HttpRequest request3 = HttpRequest.newBuilder()
-                .GET()
-                .uri(uri3)
+    @Test
+    public void deleteEpicTest() throws IOException, InterruptedException {
+        Epic epic1 = new Epic("Съездить в Москву", "обязательно до лета");
+        URI uri1 = URI.create(DEFAULT_EPICS_URI);
+        HttpRequest request1 = HttpRequest.newBuilder()
+                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(epic1)))
+                .header("Content-Type", "application/json")
+                .uri(uri1)
                 .build();
-        HttpResponse<String> response3 = client.send(request3, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response3.statusCode(), "Not right response");
+        HttpResponse<String> response1 = client.send(request1, HttpResponse.BodyHandlers.ofString());
+        Assertions.assertEquals(200, response1.statusCode(), "Not right response");
 
-        URI uri4 = URI.create(DEFAULT_PRIORITIZED_URI);
-        HttpRequest request4 = HttpRequest.newBuilder()
-                .GET()
-                .uri(uri4)
+        URI uri2 = URI.create(DEFAULT_EPICS_URI + "/1");
+        HttpRequest request2 = HttpRequest.newBuilder()
+                .DELETE()
+                .header("Content-Type", "application/json")
+                .uri(uri2)
                 .build();
-        HttpResponse<String> response4 = client.send(request4, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response4.statusCode(), "Not right response");
+        HttpResponse<String> response2 = client.send(request2, HttpResponse.BodyHandlers.ofString());
+        Assertions.assertEquals(201, response2.statusCode(), "Not right response");
     }
 }
