+++ src/ru/yandex/practicum/test/HistoryEndpointsTest.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.tests;
+package ru.yandex.practicum.test;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -7,7 +7,8 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ru.yandex.practicum.http.HttpTaskServer;
-import ru.yandex.practicum.tasks.TaskModel;
+import ru.yandex.practicum.tasks.Epic;
+import ru.yandex.practicum.tasks.Subtask;
 import ru.yandex.practicum.utils.Utils;
 
 import java.io.IOException;
@@ -17,12 +18,12 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.time.LocalDateTime;
 
-import static java.time.Month.MARCH;
+import static java.time.Month.AUGUST;
 
-public class TaskEndpointsTest {
+public class HistoryEndpointsTest {
     private HttpTaskServer httpTaskServer;
     private HttpClient client;
-    private static final String DEFAULT_TASKS_URI = "http://localhost:8080/tasks";
+    private static final String DEFAULT_HISTORY_URI = "http://localhost:8080/history";
     private static final Gson gson = new GsonBuilder()
             .registerTypeAdapter(LocalDateTime.class, new Utils.LocalDateTimeAdapter())
             .create();
@@ -40,8 +41,8 @@ public class TaskEndpointsTest {
     }
 
     @Test
-    public void getEmptyListTasksTest() throws IOException, InterruptedException {
-        URI uri = URI.create(DEFAULT_TASKS_URI);
+    public void getEmptyHistoryListTest() throws IOException, InterruptedException {
+        URI uri = URI.create(DEFAULT_HISTORY_URI);
         HttpRequest request = HttpRequest.newBuilder()
                 .GET()
                 .uri(uri)
@@ -51,47 +52,43 @@ public class TaskEndpointsTest {
     }
 
     @Test
-    public void addTaskAndGetItTest() throws IOException, InterruptedException {
-        URI uri1 = URI.create(DEFAULT_TASKS_URI);
-        TaskModel task1 = new TaskModel("Read book every day", "30 pages",
-                LocalDateTime.of(2024, MARCH, 28, 13, 0), 60);
+    public void getHistoryListTest() throws IOException, InterruptedException {
+        URI uri1 = URI.create("http://localhost:8080/epics");
+        Epic epic1 = new Epic("Съездить в Москву", "обязательно до лета");
         HttpRequest request1 = HttpRequest.newBuilder()
-                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(task1)))
+                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(epic1)))
                 .header("Content-Type", "application/json")
                 .uri(uri1)
                 .build();
         HttpResponse<String> response1 = client.send(request1, HttpResponse.BodyHandlers.ofString());
         Assertions.assertEquals(200, response1.statusCode(), "Not right response");
 
-        URI uri2 = URI.create(DEFAULT_TASKS_URI);
+        URI uri2 = URI.create("http://localhost:8080/subtasks");
+        Subtask subtask11 = new Subtask("купить билеты", "дешёвые билеты", 1,
+                LocalDateTime.of(2023, AUGUST, 28, 13, 0), 60);
         HttpRequest request2 = HttpRequest.newBuilder()
-                .GET()
+                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(subtask11)))
+                .header("Content-Type", "application/json")
                 .uri(uri2)
                 .build();
         HttpResponse<String> response2 = client.send(request2, HttpResponse.BodyHandlers.ofString());
         Assertions.assertEquals(200, response2.statusCode(), "Not right response");
-    }
 
-    @Test
-    public void deleteTaskTest() throws IOException, InterruptedException {
-        TaskModel task1 = new TaskModel("Read book every day", "30 pages",
-                LocalDateTime.of(2024, MARCH, 28, 13, 0), 60);
-        URI uri1 = URI.create(DEFAULT_TASKS_URI);
-        HttpRequest request1 = HttpRequest.newBuilder()
-                .POST(HttpRequest.BodyPublishers.ofString(gson.toJson(task1)))
-                .header("Content-Type", "application/json")
-                .uri(uri1)
-                .build();
-        HttpResponse<String> response1 = client.send(request1, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(200, response1.statusCode(), "Not right response");
 
-        URI uri2 = URI.create(DEFAULT_TASKS_URI + "/1");
-        HttpRequest request2 = HttpRequest.newBuilder()
-                .DELETE()
-                .header("Content-Type", "application/json")
-                .uri(uri2)
+        URI uri3 = URI.create("http://localhost:8080/subtasks");
+        HttpRequest request3 = HttpRequest.newBuilder()
+                .GET()
+                .uri(uri3)
                 .build();
-        HttpResponse<String> response2 = client.send(request2, HttpResponse.BodyHandlers.ofString());
-        Assertions.assertEquals(201, response2.statusCode(), "Not right response");
+        HttpResponse<String> response3 = client.send(request3, HttpResponse.BodyHandlers.ofString());
+        Assertions.assertEquals(200, response3.statusCode(), "Not right response");
+
+        URI uri4 = URI.create(DEFAULT_HISTORY_URI);
+        HttpRequest request4 = HttpRequest.newBuilder()
+                .GET()
+                .uri(uri4)
+                .build();
+        HttpResponse<String> response4 = client.send(request4, HttpResponse.BodyHandlers.ofString());
+        Assertions.assertEquals(200, response4.statusCode(), "Not right response");
     }
 }
